Auto pick of #10223 #10238 origin release calient v3.21 (#9017)
* flowlogs - we now know which side initiated the conn

* rename references to tunnel to ipip

IT is a historical artefact and we should name it properly and leave the
name tunnel for a generic overlay.

* compile distinct programs for VXLAN

We need to know that we are forwarding from and L2 overlay.

* fix pkt/bytes accounting for VXLAN

* fix pkt/bytes accounting fo IPIP as L3 dev

Pretty much we need to account for the bytes on ingress from any tunnel
device. It does not matter if it preserves the marks like vxlan or ipip
or not like wg. Simply account those bytes and packets, because it is
ingress into the system and we haven't seen this as unpacked yet for
sure.

refs https://github.com/projectcalico/calico/pull/10223